% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Github_LaTeX.R
\name{parse_latex}
\alias{parse_latex}
\title{Parse LaTeX for Markdown}
\usage{
parse_latex(rmd, new_md, git_username = "muschellij2",
  git_reponame = "Github_Markdown_LaTeX", git_branch = "master",
  text_height = 20,
  insert_string = paste0("\\n<img src=\\"\%s\%s\\" alt=\\"Equation not
  rendered\\" height=\\"", text_height, "\\">\\n"),
  raw_git_site = "https://rawgit.com", bad_string = "ZZZZZZZZZZZZZZZ")
}
\arguments{
\item{rmd}{Rmd or md file that needs to be parsed}

\item{new_md}{Output md file (usually named README.md for gitHub}

\item{git_username}{Username for repository (GitHub currently)}

\item{git_reponame}{Repository name}

\item{git_branch}{Branch of repository if not master}

\item{text_height}{Height of LaTeX rendered passed ot \code{insert_string}}

\item{insert_string}{String of HTML to put in LaTeX figure. Must have 2 \%s for sprintf}

\item{raw_git_site}{Site where to reference figure.
\code{img_prefix = file.path(raw_git_site git_username git_reponame git_branch)}.
Must reference the figures directly with proper content-type headers}

\item{bad_string}{String to sub in for dollar signs temporarily (not need to be changed)}
}
\value{
Filename of parsed MD file
}
\description{
This function will take an Rmd file and knit it, or and md file, and any
LaTex with $ or $$ delineations, create pngs of these equations, and then insert
a link to these equations to display in markdown
}
\examples{
rmd = file.path(tempdir(), "README_unparse.rmd")
download.file(
"https://raw.githubusercontent.com/muschellij2/Github_Markdown_LaTeX/master/README_unparse.rmd",
destfile = rmd, method = "curl")
new_md = file.path(tempdir(), "README.md")
parse_latex(rmd,
            new_md,
            git_username = "muschellij2",
            git_reponame = "Github_Markdown_LaTeX")
library(knitr)
new_html = pandoc(new_md, format = "html")
browseURL(new_html)

new_md = file.path(tempdir(), "README.md")
parse_latex(rmd,
            new_md,
            git_username = "muschellij2",
            git_reponame = "Github_Markdown_LaTeX")
library(knitr)
new_html = pandoc(new_md, format = "html")
browseURL(new_html)#'
}

