% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Github_LaTeX.R
\name{png_latex}
\alias{png_latex}
\title{Creates a PNG of LaTeX Equation}
\usage{
png_latex(latex, packages = c("amsmath", "amsfonts", "amssymb"), ...)
}
\arguments{
\item{latex}{single character string of LaTeX code to convert}

\item{packages}{to be loaded into LaTex using \code{\\usepackage}}

\item{...}{Arguments to be passed to \code{\link{converter}}}
}
\value{
Name of output file as a PNG
}
\description{
This code takes in LaTeX, creates a simple LaTeX document, and then
creates a PDF of this, crops it, then converts it to a PNG
}

